fix: Parse title including owner (#39)
In https://github.com/getsentry/action-prepare-release/pull/4, we unintentionally changed the format of the issue title. 

- Before: `publish: <repo>@<version>`
- After: `publish: <owner>/<repo>@<version>`

This is because the `GITHUB_REPOSITORY` env variable includes the full owner and repo name. There is also an `GITHUB_REPOSITORY_OWNER` variable, but none that only contains the repo name. This means we have two options:

1. In the prepare action, remove the owner from the repo. 
2. In the publish action, assume there's an _optional_ owner. 

In this PR, I'm going with option 2, ~~although it is not backwards compatible. This means we will have to change all existing actions before attempting the next release.~~