meta(readme): Clarify craft and publish permissions (#44)
Craft ships with Zeus as the default status and artifact providers. To work with
the release action, these have to be changed to `"github"`.

Additionally, usage instructions mixed one-time setup and repeated use. I chose
to separate them for clarity.

Finally, the expected process is that every engineer can trigger a release but
only authorized personell approve it. We clarify this by addressing every
engineer in the usage instructions while mentioning elevated permission
requirements for adding the `accepted` label.