ref(publish): Simplify publish from subfolder (#334)
This PR changes how we publish from a subfolder. Previously, we moved the contents of the subfolder to `GITHUB_WORKSPACE` as the `craft` action (image) could not work inside subfolders. This also came at the cost of losing the `.git` folder at the top level for local `git` operations.

This patch checks out the target repo under `__repo__` and then overrides the `entrypoint` and `args` of the Craft Action to run it inside the subfolder.