ci: Run publish action on production (#6627)
This change will allow us to give write access to repo once we move all the secrets to the `production` environment as we can restrict environment to the `main` branch only, preventing secret leakage on PRs. We only have the `NPM_TOKEN` moved there and will be testing this before moving over all the secrets, removing them and opening the repo up.